Retry connecting to TCP store on ECONNRESET
The retry logic dealt with ECONNREFUSED to deal with the client being
started before the server. It didn't yet deal with the server being
started but having its listen backlog exhausted. This may happen when
starting many processes that all try to connect at the same time.

The server implementation uses blocking I/O to read and write entire
messages, so it may take a bit longer to call `accept(2)` on new
connections compared to a fully event driven approach.

This commit both increases the default listen backlog on the server
side and implements retries on ECONNRESET after `connect(2)`.

ghstack-source-id: 51a1fce22b30d9c359ee6d8a5d33c8cf62324ea7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25707